service/src/questions/questions.service.ts
import { Injectable } from '@nestjs/common';
import { CreateQuestionDto } from './dto/create-question.dto';
import { UpdateQuestionDto } from './dto/update-question.dto';
import { Question } from './entities/question.entity';

@Injectable()
export class QuestionsService {
  private questions: Question[] = [];

  create(createQuestionDto: CreateQuestionDto) {
    const response = { ...createQuestionDto, id: this.questions.length };
    this.questions.push(response);
    return response;
  }

  findAll() {
    return this.questions;
  }

  findOne(id: number) {
    return `This action returns a #${id} question`;
  }

  update(id: number, updateQuestionDto: UpdateQuestionDto) {
    return `This action updates a #${id} question`;
  }

  remove(id: number) {
    return `This action removes a #${id} question`;
  }
}
